requirements.txt
flask==2.0.2
flask_cors==3.0.10

requests==2.26.0
pytz==2021.3
pandas==1.1.5
telegram-send==0.25

dnspython==2.1.0
pymongo==3.6.1